Add alias tool functionality for managing URL aliases
- Introduced a new module for the alias tool, including functions for listing and replacing aliases in 3LC objects.
- Implemented common utilities for input handling and logging to streamline alias processing.
- Enhanced the command interface to support alias listing and replacement operations, improving usability.
- Added comprehensive tests to validate the new functionalities, ensuring robust coverage for alias management scenarios.
